Huge update. MCP updater test

+++ runtime/cleanup.py
@@ -11,7 +11,6 @@ class Cleanup:
     _default_config = 'conf/mcp.cfg'
 
     def __init__(self, conffile=None):
-        print(os.getcwd())
         self.conffile = conffile
         self.readconf()
         self.confdir = self.config.get("DEFAULT", "DirConf")
@@ -29,21 +28,24 @@ class Cleanup:
             self.systemext = "sh"
 
     def start(self):
-        """
-        ARE YOU *REALLY* SURE?!
-        :return:
-        """
         print("> Welcome to the LTS cleanup script!")
         print("> This script will delete your workspace and set most of it to factory defaults.")
-        print("> Are you sure you want to clean up your workspace? [y/N]")
+        print("> Input 's' if you want to only clear the source and bin folders")
+        print("> Are you sure you want to clean up your workspace? [y/N/s]")
         inp = input(": ")
-        if inp.lower() != "y":
-            sys.exit(1)
+        clearsrc = False
+        if inp.lower() == "s":
+            clearsrc = True
+        if inp.lower() != "y" and inp.lower() != "s":
+            return
         print("> Are you *REALLY* sure you want to clean up your workspace? [y/N]")
-        print("> This deletes ALL your source files and jars! This is NOT recoverable!")
+        if clearsrc:
+            print("> This deletes ALL your source files!")
+        else:
+            print("> This deletes ALL your source files and jars! This is NOT recoverable!")
         inp = input(": ")
         if inp.lower() != "y":
-            sys.exit(1)
+            return
 
         print("> Commencing the purge of the universe...")
         no_error = True
@@ -51,22 +53,24 @@ class Cleanup:
         print("> Deleting \"" + self.jardir + "\"...")
         deltime = time.time()
         # Delete jars while keeping server.properties.
-        try:
-            if os.path.exists(self.jardir):
-                if not os.path.exists(self.tempdir):
-                    os.makedirs(self.tempdir)
-                if os.path.exists(os.path.join(self.jardir, "server.properties")):
-                    shutil.copy2(os.path.join(self.jardir, "server.properties"), self.tempdir)
-                shutil.rmtree(self.jardir)
-                os.makedirs(self.jardir)
-                if os.path.exists(os.path.join(self.tempdir, "server.properties")):
-                    shutil.copy2(os.path.join(self.tempdir, "server.properties"), self.jardir)
-        except Exception as e:
-            no_error = False
-            print("> Couldn't clear \"" + self.jardir + "\"!")
-            traceback.print_exc()
+        if not clearsrc:
+            try:
+                if os.path.exists(self.jardir):
+                    if not os.path.exists(self.tempdir):
+                        os.makedirs(self.tempdir)
+                    if os.path.exists(os.path.join(self.jardir, "server.properties")):
+                        shutil.copy2(os.path.join(self.jardir, "server.properties"), self.tempdir)
+                    shutil.rmtree(self.jardir)
+                    os.makedirs(self.jardir)
+                    if os.path.exists(os.path.join(self.tempdir, "server.properties")):
+                        shutil.copy2(os.path.join(self.tempdir, "server.properties"), self.jardir)
+            except Exception as e:
+                no_error = False
+                print("> Couldn't clear \"" + self.jardir + "\"!")
+                traceback.print_exc()
 
-        for dir in [self.reobfdir, self.bindir, self.srcdir, self.tempdir]:
+        cleardirs = [self.reobfdir, self.bindir, self.srcdir, self.tempdir]
+        for dir in cleardirs:
             print("> Deleting \"" + dir + "\"...")
             try:
                 if os.path.exists(dir):
@@ -76,32 +80,33 @@ class Cleanup:
                 print("> Couldn't clear \"" + dir + "\"!")
                 traceback.print_exc()
 
-        print("> Deleting non-default config...")
-        try:
-            if os.path.exists(self.confdir):
-                if os.path.exists(os.path.join(self.confdir, "patches")) and os.path.isdir(os.path.join(self.confdir, "patches")):
-                    shutil.rmtree(os.path.join(self.confdir, "patches"))
-                for file in os.listdir(self.confdir):
-                    if os.path.isfile(os.path.join(self.confdir, file)) and file not in ["mcp.cfg"]:
-                        os.unlink(os.path.join(self.confdir, file))
-        except Exception as e:
-            no_error = False
-            print("> Couldn't clear \"" + self.confdir + "\"!")
-            traceback.print_exc()
+        if not clearsrc:
+            print("> Deleting non-default config...")
+            try:
+                if os.path.exists(self.confdir):
+                    if os.path.exists(os.path.join(self.confdir, "patches")) and os.path.isdir(os.path.join(self.confdir, "patches")):
+                        shutil.rmtree(os.path.join(self.confdir, "patches"))
+                    for file in os.listdir(self.confdir):
+                        if os.path.isfile(os.path.join(self.confdir, file)) and file not in ["mcp.cfg"]:
+                            os.unlink(os.path.join(self.confdir, file))
+            except Exception as e:
+                no_error = False
+                print("> Couldn't clear \"" + self.confdir + "\"!")
+                traceback.print_exc()
 
-        print("> Deleting system specific files from root...")
-        try:
-            for file in ["decompile", "recompile", "reobfuscate", "startclient", "startserver", "updatemcp", "updatemd5"]:
-                if os.path.exists(file + "." + self.systemext) and os.path.isfile(file + "." + self.systemext):
-                    os.unlink(file + "." + self.systemext)
-        except Exception as e:
-            no_error = False
-            print("> Couldn't clear system specific files!")
-            traceback.print_exc()
+            print("> Deleting system specific files from root...")
+            try:
+                for file in ["decompile", "recompile", "reobfuscate", "startclient", "startserver", "updatemcp", "updatemd5"]:
+                    if os.path.exists(file + "." + self.systemext) and os.path.isfile(file + "." + self.systemext):
+                        os.unlink(file + "." + self.systemext)
+            except Exception as e:
+                no_error = False
+                print("> Couldn't clear system specific files!")
+                traceback.print_exc()
 
-        if no_error:
+        if no_error and not clearsrc:
             os.unlink("cleanup." + self.systemext)
-        else:
+        elif not no_error:
             print("> Cleanup file has not been deleted because an error occurred earlier.")
         print('> Done in %.2f seconds' % (time.time() - deltime))
 
@@ -125,7 +130,5 @@ if __name__ == '__main__':
     cleanup = Cleanup()
     try:
         cleanup.start()
-        sys.exit(0)
     except:
         traceback.print_exc()
-        sys.exit(1)
